requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
crispy-bootstrap5==0.6
Django==4.0.4
django-autoslug==1.9.8
django-ckeditor==6.4.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-uuslug==2.0.0
idna==3.3
Pillow==9.1.0
psycopg2==2.9.3
python-slugify==6.1.2
requests==2.27.1
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.9
